Register Slahs commands
