+++ github/app.json
@@ -1,7 +1,7 @@
 {
     "id": "826f0d95-9e25-48a6-a781-a32f147230a5",
     "version": "0.0.1",
-    "requiredApiVersion": "^1.4.0",
+    "requiredApiVersion": "^1.19.0",
     "iconFile": "icon.png",
     "author": {
         "name": "Samad Yar Khan",
@@ -11,5 +11,6 @@
     "name": "Github",
     "nameSlug": "github",
     "classFile": "GithubApp.ts",
-    "description": "The ultimate app extending Rocket.Chat for all developers collaborating on Github"
+    "description": "The ultimate app extending Rocket.Chat for all developers collaborating on Github",
+    "implements": []
 }
